refacor : HttpRequest 구조체 생성 및 기존 로직 개선
* 기존 IOUtils, HttpRequestUtils에 존재하는 메서드를 활용하여 요청 상태줄, 헤더 정보를 저장하는 HttpRequest 구조체를 별도 생성하였습니다.
* 별도로 사용하지 않는 IOUtils, HttpRequestUtils 관련 클래스 파일 및 패키지는 삭제했습니다.
