--- src/test/java/util/HttpRequestUtilsTest.java
@@ -1,89 +1,0 @@
-package util;
-
-import static org.assertj.core.api.Assertions.*;
-
-import java.util.Map;
-
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import util.HttpRequestUtils.Pair;
-
-public class HttpRequestUtilsTest {
-    @Test
-    public void parseQueryString() {
-        String queryString = "userId=javajigi";
-        Map<String, String> parameters = HttpRequestUtils.parseQueryString(queryString);
-        assertThat(parameters.get("userId")).isEqualTo("javajigi");
-        assertThat(parameters.get("password")).isNull();
-
-        queryString = "userId=javajigi&password=password2";
-        parameters = HttpRequestUtils.parseQueryString(queryString);
-        assertThat(parameters.get("userId")).isEqualTo("javajigi");
-        assertThat(parameters.get("password")).isEqualTo("password2");
-    }
-
-    @Test
-    public void parseQueryString_null() {
-        Map<String, String> parameters = HttpRequestUtils.parseQueryString(null);
-        assertThat(parameters.isEmpty()).isTrue();
-
-        parameters = HttpRequestUtils.parseQueryString("");
-        assertThat(parameters.isEmpty()).isTrue();
-
-        parameters = HttpRequestUtils.parseQueryString(" ");
-        assertThat(parameters.isEmpty()).isTrue();
-    }
-
-    @Test
-    public void parseQueryString_invalid() {
-        String queryString = "userId=javajigi&password";
-        Map<String, String> parameters = HttpRequestUtils.parseQueryString(queryString);
-        assertThat(parameters.get("userId")).isEqualTo("javajigi");
-        assertThat(parameters.get("password")).isNull();
-    }
-
-    @Test
-    public void parseCookies() {
-        String cookies = "logined=true; JSessionId=1234";
-        Map<String, String> parameters = HttpRequestUtils.parseCookies(cookies);
-        assertThat(parameters.get("logined")).isEqualTo("true");
-        assertThat(parameters.get("JSessionId")).isEqualTo("1234");
-        assertThat(parameters.get("session")).isNull();
-    }
-
-    @Test
-    public void getKeyValue() throws Exception {
-        Pair pair = HttpRequestUtils.getKeyValue("userId=javajigi", "=");
-        assertThat(pair).isEqualTo(new Pair("userId", "javajigi"));
-    }
-
-    @Test
-    public void getKeyValue_invalid() throws Exception {
-        Pair pair = HttpRequestUtils.getKeyValue("userId", "=");
-        assertThat(pair).isNull();
-    }
-
-    @Test
-    public void parseHeader() throws Exception {
-        String header = "Content-Length: 59";
-        Pair pair = HttpRequestUtils.parseHeader(header);
-        assertThat(pair).isEqualTo(new Pair("Content-Length", "59"));
-    }
-
-    @Test
-    @DisplayName("두 페어가 가진 문자열이 일치하면 true, 다르면 false가 나와야 한다.")
-    public void testPairEquals() {
-        Pair aa = new Pair("a", "a");
-        Pair AA = new Pair("a", "a");
-        Pair ab = new Pair("a", "b");
-        Pair ba = new Pair("b", "a");
-        Pair bb = new Pair("b", "b");
-
-        assertThat(aa.equals(aa)).isTrue();
-        assertThat(aa.equals(AA)).isTrue();
-        assertThat(aa.equals(ab)).isFalse();
-        assertThat(aa.equals(ba)).isFalse();
-        assertThat(aa.equals(bb)).isFalse();
-    }
-}
